test(core): split fact kernel characterization responsibilities
